城市的最短路径开发

+++ src/RA.java
@@ -9,7 +9,7 @@ public class RA {
 	 * @param y, y coordinate
 	 * @param radius, the radius that the restricted area has
 	 */
-	RA(int x, int y, int radius) {
+	RA(double x, double y, double radius) {
 		this.x = x;
 		this.y = y;
 		this.radius = radius;
